updated first two tests and added garbage collection to processor
