+++ tests/test_dem_generation_local_1/test_dem_generation_local_1.py
@@ -17,6 +17,8 @@ import shapely
 import pdal
 import logging
 import gc
+import pytest
+import sys
 
 from src.geofabrics import processor
 
@@ -211,7 +213,64 @@ class Test(unittest.TestCase):
                         shutil.rmtree(file)
                 shutil.rmtree(path)
 
-    def test_result_dem(self):
+    @pytest.mark.skipif(sys.platform != "win32", reason="Windows test - this is less strict")
+    def test_result_dem_windows(self):
+        """A basic comparison between the generated and benchmark DEM"""
+  
+        decimal_tolerance=5
+
+        # Load in benchmark DEM
+        file_path = self.cache_dir / self.instructions["data_paths"]["benchmark_dem"]
+        with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as benchmark_dem:
+            benchmark_dem = benchmark_dem.squeeze("band", drop=True)
+        # Load in result DEM
+        file_path = self.results_dir / self.instructions["data_paths"]["result_dem"]
+        with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as test_dem:
+            test_dem = test_dem.squeeze("band", drop=True)
+        # Compare DEMs z - load both from file as rioxarray.rioxarray.open_rasterio
+        # ignores index order
+        diff_array = test_dem.z.data - benchmark_dem.z.data
+        logging.info(f"DEM z array diff is: {diff_array[diff_array != 0]}")
+        numpy.testing.assert_array_almost_equal(
+            test_dem.z.data,
+            benchmark_dem.z.data,
+            decimal=decimal_tolerance,
+            err_msg="The generated result_dem z has different data from the "
+            "benchmark_dem",
+        )
+
+        # Compare DEMs data source classification
+        diff_array = test_dem.data_source.data - benchmark_dem.data_source.data
+        logging.info(f"DEM data source diff is: {diff_array[diff_array != 0]}")
+        numpy.testing.assert_array_almost_equal(
+            test_dem.data_source.data,
+            benchmark_dem.data_source.data,
+            decimal=decimal_tolerance,
+            err_msg="The generated test data_source layer has different data "
+            "from the benchmark",
+        )
+
+        # Compare DEMs lidar source classification
+        diff_array = test_dem.lidar_source.data - benchmark_dem.lidar_source.data
+        logging.info(f"DEM LiDAR source diff is: {diff_array[diff_array != 0]}")
+        numpy.testing.assert_array_almost_equal(
+            test_dem.lidar_source.data,
+            benchmark_dem.lidar_source.data,
+            decimal=decimal_tolerance,
+            err_msg="The generated test lidar_source layer has different data "
+            "from the benchmark",
+        )
+
+        # explicitly free memory as xarray seems to be hanging onto memory
+        test_dem.close()
+        benchmark_dem.close()
+        del test_dem
+        del benchmark_dem
+
+    @pytest.mark.skipif(
+        sys.platform != "linux", reason="Linux test - this is more strict"
+    )
+    def test_result_dem_linux(self):
         """A basic comparison between the generated and benchmark DEM"""
 
         # Load in benchmark DEM
@@ -258,7 +317,6 @@ class Test(unittest.TestCase):
         benchmark_dem.close()
         del test_dem
         del benchmark_dem
-        gc.collect()
 
 
 if __name__ == "__main__":
